feat: Add durability statistic to character items and update related systems
- Introduced a new durability statistic for character items, allowing for tracking and management of item durability.
- Updated CharacterData to handle durability in statistics calculations.
- Enhanced CharacterPlayer to check durability before using items.
- Modified CharacterPlayerHud to display item durability in the UI.
- Implemented camera radius adjustment functionality in CharacterPlayerCamera.
- Updated input actions to include SetCameraRadius for camera control.
- Refactored inventory management to accommodate durability checks and updates.
- Adjusted item creation in ItemsDBSO to initialize durability values.
- Improved utility functions to refresh consumable displays based on durability status.

+++ Assets/Scripts/Navigation/CharacterGridNavigator.cs
@@ -208,14 +208,14 @@ public class CharacterGridNavigator : CharacterMovementBase
 
     int ResolveJumpDistance()
     {
-        return characterBase.charactersData[characterBase.characterIndex]
+        return (int)characterBase.charactersData[characterBase.characterIndex]
                             .statistics[CharacterData.TypeStatistic.JumpDistance]
                             .currentValue;
     }
 
     int ResolveDropDistance()
     {
-        return characterBase.charactersData[characterBase.characterIndex]
+        return (int)characterBase.charactersData[characterBase.characterIndex]
                             .statistics[CharacterData.TypeStatistic.DropDistance]
                             .currentValue;
     }
